'class' attribute for all components
The 'class' attribute in the XML layout spec is now observed by all components.

The attribute is not a parameter to the component constructor anymore. Instead it is directly
taken from the DOMElement which is a parameter to the component constructor itself, making the
additional 'class' parameter redundent except in special cases. For these special cases an
addClass and a setClass method were added to the Component class.

Change-Id: I9034b488a67c0b061045c066c4b7ec251dc7cf1d