Add 'uselayout' URL parameter
To select a specific layout different from the one defined in
`$egChameleonLayoutFile` you can add the `uselayout` parameter to the URL.
However for some wikis it might not be desirable to have this feature. So to
make this work you have to define the available layouts in
LocalSettings.php. E.g. to include all layouts delivered with Chameleon add
```php
$egChameleonAvailableLayoutFiles = array(
    'standard'   => __DIR__ . '/skins/chameleon/layouts/standard.xml',
    'navhead'    => __DIR__ . '/skins/chameleon/layouts/navhead.xml',
    'fixedhead'  => __DIR__ . '/skins/chameleon/layouts/fixedhead.xml',
    'stickyhead' => __DIR__ . '/skins/chameleon/layouts/stickyhead.xml',
    'clean'      => __DIR__ . '/skins/chameleon/layouts/clean.xml',
);
```

Change-Id: Ie1d8eac82424053e5490d23d9ad864d84f725481