PersonalTools: Add attribute hideNewtalkNotifier + use it for Standard layout
The attribute was introduced to keep backwards compatibility. If the
PersonalTools component is used, it is recommended to always set this attribute
to yes and use an independent NewtalkNotifier component.

The standard layout uses this now.

hideNewtalkNotifier:
* Already deprecated.
* Allowed values: Boolean (yes|no)
* Default: no
* Optional.

If set the newtalk notifier will not be shown.
This attribute has no effect when used inside the NavbarHorizontal component.