legal: Remove misleading/incorrect text about bundling
MediaWiki is licensed as GPL v2 or later, so including any GPL v3 projects
is totally OK, since you're just using it under the terms of v3. And since
MediaWiki depends on and bundles Apache 2.0 projects (which is only
compatible with GPL v3), MediaWiki itself is effectively GPL v3. So there's
no problem with bundling GPL v3 projects with MediaWiki, and in fact the
1.31 release will contain GPL v3 or later code in the tarball.