Fix #91: Ensure mediawiki.skinning.content module is loadable
In MW 1.33 the ResourceLoaderRegisterModules hook is called before
the standard MW module definitions are loaded from Resources.php.
It is therefore not possible to work with these module definitions
from inside the hook.

It is in particular not possible to register a module
0.mediawiki.skinning.content based on the module
mediawiki.skinning.content.

Instead of registering 0.mediawiki.skinning.content to enforce the
correct loading order for modules a module zzz.ext.bootstrap.styles
is registered.