Add option to display username-only in PersonalTools (fixes #343)
This adds a `username-only` option to the `showUserName` attribute in the
`PersonalTools` component (when used within the `NavbarHorizontal`), to show
*only* the username for logged-in users.  The prior `yes` option has been
renamed to `try-realname` for clarity, and the `no` option has been renamed
to `none` (the new default).

For backwards compatibility, `no` and `yes` are still accepted, but a
deprecation note has been added to the code.