Move 'title' attribute of NavMenu toggles to the <a> elements
Currently, the 'title' attribute on a NavMenu section/item is placed on the
top-level wrapper <div> element, instead of the dropdown-toggle's <a>
element.  Everywhere else, 'title' attributes are placed directly on the
<a> element of the button/item/link.

Since the top-level wrapper <div> encompasses the menu as well as the toggle,
any items in the menu that do not have their own explicit 'title' attribute
will inherit the top-level title.  In other words, menu items that should
not have tooltips will get the same tooltip as the dropdown-toggle.

This commit moves the 'title' attribute in NavMenu off the top-level <div>
and onto the toggle <a> element.