Tweak CSS to fix grouping of "right" items when navbar is collapsed
It is not clear what the `.flex-row` on `.navbar-nav.right` elements
was intended to accomplish, but the effect that it has is that, when
the navbar is collapsed and switches to a vertical layout, all the
right-grouped navbar items are jammed together into a single row
(instead of switching to a vertical/column layout like the non-right-
-grouped items).  Removing the `.flex-row` fixes that.

This commit also removes the `.justify-content-center`, which does
not seem to have any effect at all on any layout/orientation/screen.
(Since the `.navbar-nav.right` wrapper is always squeezed to the right,
there is no free space into which its children can be centered.)

For what it's worth, this commit does leave the `.mt-4` in place;
it does not seem to have any effect in the horizontal layout, and in
the collapsed/vertical layout it appears to introduce a gap between
the "left" grouped (top) items and the "right" grouped (bottom) items.