Add a hook for customizing the PersonalTools toggle <a> inner html
This commit adds a `ChameleonNavbarHorizontalPersonalToolsLinkInnerHtml`
hook which can be used to completely customize the inner html of the
dropdown toggle link of `PersonalTools`.

This hook will be passed three parameters:
 * `$newtalkNotifierHtml` (by reference)
 * `$userNameHtml` (by reference)
 * the `PersonalTools` object

The first two parameters are strings bearing the values computed by
the PersonalTools component (according to its configuration, etc),
which the hook function may modify (or leave alone).

After the hook function is complete, the values of `$newtalkNotifierHtml`
and `$userNameHtml` will be concatenated (in that order) to yield the
raw html content of the dropdown toggle's `<a>` element.