Add option to PersonalTools to show lone menu items directly in navbar
This commit adds a `promoteLoneItems` parameter to the `PersonalTools`
component.  If there is only one personal tool, and its name is in the
parameter, then the link for that tool will be rendered directly in
the navbar (instead of in a dropdown menu).

A prime use of this parameter is for a semi-private wiki, where a
logged-out user only has the `login` tool available.  With this
parameter, the login button can be directly rendered (and thus, clicked)
in the navbar, instead of being hidden in a menu activated by a dropdown
toggle.