Declare skin to be responsive, remove extra viewport tag (fixes #470)
 * Tell the MW core that this skin supports responsive behavior.
 * Stop manually emitting a "viewport" meta tag from Chameleon.

MW `Skin` always emits a "viewport" meta tag --- but it needs to know if
a skin supports responsive behavior in order to emit the correct tag.

On top of that, the additional tag which `Chameleon` was emitting was either
redundant, or had different/competing values for `width` (so some mobile
browsers still performed non-responsive rendering, despite the extra tag).